# docs/swagger.yaml
definitions:
...
  application.UpdateStatusRequest:
    properties:
      notes:
        type: string
      status:
        type: string
    required:
    - status
    type: object
...
info:
  contact: {}
paths:
...
  /api/applications/{applicationId}:
    delete:
      description: Delete a job application (student only)
      parameters:
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      responses:
        "200":
          description: Application removed successfully
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Could not remove application
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove Application
      tags:
      - Applications
      x-swagger-ui: true
    get:
      consumes:
      - application/json
      description: Get a job application by its ID (student or employer)
      parameters:
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Application retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Application By ID
      tags:
      - Applications
      x-swagger-ui: true
  /api/applications/{applicationId}/status:
    put:
      description: Update the status of a job application (student only)
      parameters:
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      - description: Status update request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/application.UpdateStatusRequest'
      responses:
        "200":
          description: Application status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Application Status
      tags:
      - Applications
      x-swagger-ui: true
  /api/applications/my:
    get:
      consumes:
      - application/json
      description: Get all applications submitted by the current student
      produces:
      - application/json
      responses:
        "200":
          description: Applications fetched successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get My Applications
      tags:
      - Applications
...
  /api/jobs/{id}/applications:
    get:
      consumes:
      - application/json
      description: Get all applications submitted for a specific job (employer only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Applications fetched successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Applications By Job
      tags:
      - Applications
      x-swagger-ui: true
  /api/jobs/{id}/applications/{applicationId}/status:
    put:
      description: Update the status of a job application (employer only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Application ID
        in: path
        name: applicationId
        required: true
        type: string
      - description: Status update request
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/application.UpdateStatusRequest'
      responses:
        "200":
          description: Application status updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Application Status by Employer
      tags:
      - Applications
      x-swagger-ui: true
  /api/jobs/{id}/apply:
    post:
      consumes:
      - multipart/form-data
      description: Apply for a job with resume upload (student only)
      parameters:
      - description: Job ID
        in: path
        name: id
        required: true
        type: string
      - description: Cover letter text
        in: formData
        name: cover_letter
        type: string
      - description: Resume file (PDF, DOC, DOCX, max 10MB)
        in: formData
        name: resume_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Application submitted successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or file format
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Permission denied
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Apply for Job
      tags:
      - Applications
...
swagger: "2.0"